Updated PumpControl, initializer, tostring and notify methods. Also contains outcommented code to recieve from the overlord.
Updated initializer to take name as pinname has outcommented overlord
command to register itself. Has tostring method which should return the
name.

Notify method written to take a "Pump ON" or a "Pump OFF" command, which
should either turn the pump on or off accordingly. if it recieves any
other command it should return a  "Unknown command".